requirements.txt
beir==2.0.0
faiss-cpu
faiss-gpu
nltk==3.8.1
numpy==1.26.4
openai==1.60.0
pandas==2.2.2
rank_bm25==0.2.2
scikit_learn==1.2.2
sentence_transformers==2.6.1
torch==2.2.2
tqdm==4.65.0
transformers==4.39.3
tenacity>=8.0
fastapi[standard]